#2388: Single core version is fine again

+++ tt_metal/kernels/dataflow/writer_max_pool_2d_single_core.cpp
@@ -23,15 +23,15 @@ void kernel_main() {
     const uint32_t nbatch = get_arg_val<uint32_t>(27);
     const uint32_t out_hw = get_arg_val<uint32_t>(29);
     const uint32_t start_out_h_i = get_arg_val<uint32_t>(30);
-    const int32_t end_out_h_i = get_arg_val<int32_t>(31);
+    const uint32_t end_out_h_i = get_arg_val<uint32_t>(31);
 
     constexpr bool is_out_dram = get_compile_time_arg_val(1) == 1;
     constexpr uint32_t out_nelems = get_compile_time_arg_val(3);
 
     constexpr uint32_t out_cb_id = tt::CB::c_out0;
 
-    DPRINT << "start_out_h_i: " << start_out_h_i << ENDL();
-    DPRINT << "end_out_h_i: " << end_out_h_i << ENDL();
+    DPRINT << "start_out_h_i: " << (uint) start_out_h_i << ENDL();
+    DPRINT << "end_out_h_i: " << (uint) end_out_h_i << ENDL();
 
     // ROW_MAJOR output
     const InterleavedAddrGen<is_out_dram> s_out = {
@@ -43,7 +43,7 @@ void kernel_main() {
     for (uint32_t batch = 0; batch < nbatch; ++ batch) {
         uint32_t out_row_id = start_out_h_i;
         // for every output pixel
-        for (int32_t out_h_i = start_out_h_i; out_h_i < end_out_h_i; ++ out_h_i) {
+        for (uint32_t out_h_i = start_out_h_i; out_h_i < end_out_h_i; ++ out_h_i) {
             for (uint32_t out_w_i = 0; out_w_i < out_w_loop_count; ++ out_w_i) {
                 cb_wait_front(out_cb_id, out_nelems * 2);
                 // kernel_profiler::mark_time(13);
